Refactor: Modularize asset optimization components
