+++ fp-performance-suite/.modularization-manifest
@@ -1,0 +1,85 @@
+# Modularization Manifest
+# Generated: 2025-10-07
+# Plugin: FP Performance Suite v1.0.1
+
+## NEW FILES CREATED (8 classes + 3 docs)
+
+### Core Modules
+src/Services/Assets/HtmlMinifier.php              # 1.5K - HTML minification
+src/Services/Assets/ScriptOptimizer.php           # 1.9K - Script tag optimization
+src/Services/Assets/WordPressOptimizer.php        # 1.8K - WordPress core optimizations
+
+### Combiners Package
+src/Services/Assets/Combiners/DependencyResolver.php    # 4.4K - Topological sort
+src/Services/Assets/Combiners/AssetCombinerBase.php     # 7.4K - Base combiner class
+src/Services/Assets/Combiners/CssCombiner.php           # 5.0K - CSS combination
+src/Services/Assets/Combiners/JsCombiner.php            # 6.5K - JS combination
+
+### Resource Hints
+src/Services/Assets/ResourceHints/ResourceHintsManager.php  # 6.6K - DNS prefetch & preload
+
+### Documentation
+docs/MODULARIZATION_REPORT.md          # Complete modularization report
+MODULARIZATION_CHANGELOG.md            # Detailed changelog
+src/Services/Assets/README.md          # Module usage guide
+
+## MODIFIED FILES (2)
+
+src/Services/Assets/Optimizer.php      # Refactored from 944 to 370 lines (-61%)
+src/Plugin.php                         # Updated ServiceContainer registrations
+
+## DIRECTORIES CREATED (2)
+
+src/Services/Assets/Combiners/         # Asset combination logic
+src/Services/Assets/ResourceHints/     # Resource hints management
+
+## METRICS
+
+Before:
+- Optimizer.php: 944 lines
+- Asset classes: 2
+- Max file size: 944 lines
+
+After:
+- Optimizer.php: 370 lines (-61%)
+- Asset classes: 10 (+400%)
+- Max file size: ~250 lines (-73%)
+- Total new files: 11 (8 classes + 3 docs)
+
+## ARCHITECTURE IMPROVEMENTS
+
+✓ Single Responsibility Principle applied
+✓ Dependency Injection implemented
+✓ Full backward compatibility maintained
+✓ Zero breaking changes
+✓ Improved testability
+✓ Enhanced maintainability
+✓ Better code organization
+
+## BACKWARD COMPATIBILITY
+
+All deprecated methods still work:
+- Optimizer::minifyHtml()           → HtmlMinifier::minify()
+- Optimizer::dnsPrefetch()          → ResourceHintsManager::addDnsPrefetch()
+- Optimizer::preloadResources()     → ResourceHintsManager::addPreloadHints()
+- Optimizer::heartbeatSettings()    → WordPressOptimizer::configureHeartbeat()
+
+## VERIFICATION
+
+To verify the modularization:
+1. Check all files exist: ls -la src/Services/Assets/
+2. Verify syntax: php -l src/Services/Assets/*.php
+3. Run tests: phpunit tests/
+4. Check backwards compat: Review deprecated method calls
+
+## NEXT STEPS
+
+1. [ ] Modularize WebPConverter.php (506 lines)
+2. [ ] Add unit tests for new modules
+3. [ ] Create formal interfaces (CombinerInterface)
+4. [ ] Document advanced usage examples
+
+---
+Status: COMPLETED ✓
+Date: 2025-10-07
+Author: Francesco Passeri
